refactor(minecraft): extract asset management into ModAssets
